GradeStatus concern for Grade and ChallengeGrade
